+++ Tempest/dx/directx11.h
@@ -126,7 +126,7 @@ class DirectX11 : public AbstractAPI {
 
     void bindVertexBuffer( AbstractAPI::Device *d,
                            AbstractAPI::VertexBuffer*,
-                           int vsize  ) const;
+                           size_t vsize  ) const;
     void bindIndexBuffer( AbstractAPI::Device *d,
                           AbstractAPI::IndexBuffer* ) const;
