+++ Tempest/dx/directx9.h
@@ -116,7 +116,7 @@ class DirectX9 : public AbstractAPI {
 
     void bindVertexBuffer( AbstractAPI::Device *d,
                            AbstractAPI::VertexBuffer*,
-                           int vsize  ) const;
+                           size_t vsize  ) const;
     void bindIndexBuffer( AbstractAPI::Device *d,
                           AbstractAPI::IndexBuffer* ) const;
